# A2reservasREST/openapi.json
{"openapi":"3.0.2","info":{"title":"A2reservasREST","description":"Gestión de reservas. Ingeniería web","version":"0.1.0"},"paths":{"/booking":{"get":{"tags":["Booking"],"summary":"Get Bookings","description":"Returns a list of bookings that match the given criteria. 10 bookings max","operationId":"get_bookings","parameters":[{"required":false,"schema":{"title":"User Id","type":"string"},"name":"user_id","in":"query"},{"required":false,"schema":{"title":"Before Date","type":"string","format":"date"},"name":"before_date","in":"query"},{"required":false,"schema":{"title":"After Date","type":"string","format":"date"},"name":"after_date","in":"query"},{"required":false,"schema":{"title":"Skip","exclusiveMinimum":0.0,"type":"integer"},"name":"skip","in":"query"},{"required":false,"schema":{"$ref":"#/components/schemas/SortBookingEnum"},"name":"sort_by","in":"query"},{"required":false,"schema":{"title":"Ascending","type":"boolean","default":false},"name":"ascending","in":"query"},{"required":false,"schema":{"title":"House Id","type":"string"},"name":"house_id","in":"query"},{"required":false,"schema":{"title":"Owner Id","type":"string"},"name":"owner_id","in":"query"},{"required":false,"schema":{"$ref":"#/components/schemas/BookingStateEnum"},"name":"state","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Bookings","type":"array","items":{"$ref":"#/components/schemas/Booking"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Booking"],"summary":"Create Booking","description":"Creates a new booking","operationId":"new_booking","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewBooking"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaypalCreateOrderRequestBody"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/booking/{booking_id}":{"get":{"tags":["Booking"],"summary":"Get Booking By Id","description":"Returns all the information related to the given `booking_id`","operationId":"get_booking_by_id","parameters":[{"required":true,"schema":{"title":"Booking Id","type":"string"},"name":"booking_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Booking"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Booking"],"summary":"Update Booking","description":"Updates an existing booking","operationId":"update_booking","parameters":[{"required":true,"schema":{"title":"Booking Id","type":"string"},"name":"booking_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBooking"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Booking"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Booking"],"summary":"Cancel Booking","description":"Cancels the booking identified by the given `booking_id`","operationId":"cancel_booking","parameters":[{"required":true,"schema":{"title":"Booking Id","type":"string"},"name":"booking_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ping":{"get":{"tags":["Booking"],"summary":"Ping","description":"Responds with an \"Available\" message, if the service is available","operationId":"ping_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}}}}}},"components":{"schemas":{"ApiError":{"title":"ApiError","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","type":"string"}},"description":"An error response"},"Booking":{"title":"Booking","required":["id","house_id","user_id","start_date","end_date"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"house_id":{"title":"House Id","type":"string"},"user_id":{"title":"User Id","type":"string"},"start_date":{"title":"Start Date","type":"string","format":"date"},"end_date":{"title":"End Date","type":"string","format":"date"},"state":{"allOf":[{"$ref":"#/components/schemas/BookingStateEnum"}],"default":"reserved"},"paypal_transaction_id":{"title":"Paypal Transaction Id","type":"string"}},"description":"Contains all the known information about a booking"},"BookingStateEnum":{"title":"BookingStateEnum","enum":["reserved","canceled"],"type":"string","description":"An enumeration."},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Message":{"title":"Message","required":["message"],"type":"object","properties":{"message":{"title":"Message","type":"string"}},"description":"A single message payload"},"NewBooking":{"title":"NewBooking","required":["house_id","user_id","start_date","end_date"],"type":"object","properties":{"house_id":{"title":"House Id","type":"string"},"user_id":{"title":"User Id","type":"string"},"start_date":{"title":"Start Date","type":"string","format":"date"},"end_date":{"title":"End Date","type":"string","format":"date"}},"description":"Payload for creating new bookings"},"PaypalAmount":{"title":"PaypalAmount","required":["value"],"type":"object","properties":{"value":{"title":"Value","type":"string"}}},"PaypalCreateOrderRequestBody":{"title":"PaypalCreateOrderRequestBody","required":["purchase_units"],"type":"object","properties":{"purchase_units":{"title":"Purchase Units","type":"array","items":{"$ref":"#/components/schemas/PaypalPurchaseUnit"}}}},"PaypalPurchaseUnit":{"title":"PaypalPurchaseUnit","required":["amount"],"type":"object","properties":{"amount":{"$ref":"#/components/schemas/PaypalAmount"}}},"SortBookingEnum":{"title":"SortBookingEnum","enum":["start_date","end_date"],"type":"string","description":"An enumeration."},"UpdateBooking":{"title":"UpdateBooking","required":["user_id","paypal_transaction_id"],"type":"object","properties":{"user_id":{"title":"User Id","type":"string"},"paypal_transaction_id":{"title":"Paypal Transaction Id","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}},"servers":[{"url":"/","description":"Default"},{"url":"http://localhost:8001","description":"Docker compose"}]}